author-info-blocks: automate png generation of sample doc
A PDF document is generated from the sample Markdown file and then
converted to PNG via ImageMagick.

The dependency used for PDF conversions seems to be a security nightmare
when used on untrusted input, it is disabled on most systems but must be
enabled for the make targets to build correctly.