table-short-captions: ensure test do run
The new tests introduced by 5466859 (table-short-captions: replace
golden files with expected strings, 2019-05-10) have been broken all
this time. They always pass because they actually do not check anything.

The assert_contains function receives the test string to be searched
on stdin, but inside the function it is wrongly assumed to be in $1,
which is not set. Thus, grep searches for the empty string and all the
tests always pass.

Fix by consuming stdin inside the function to get the content to check,
which ensures the content is given as the query to grep.